refactor: report header 컴포넌트 분리
